Condense Eio__core into Eio__thin
We focus on fibers that return promises.
We don't use switches and the cancellation context hierarchy. Instead
each fiber returns a cancellation context when it is forked.
This keeps the basic feature of cancellation but makes it harder to
use. (also no resource cleanup in switches but we don't support
OS resources here either.)

If we assume that the Cells module is already verified, the initial
version of the thinned out modules only has 291 lines of code. Though it
might need to grow a bit to get a really usable/bugless version.

$ cloc lib_eio/thin/{broadcast.ml,cancel.ml,fiber.ml,promise.ml,suspend.ml} lib_eio/mock/thin_backend.ml

-------------------------------------------------------------------------------
Language                     files          blank        comment           code
-------------------------------------------------------------------------------
OCaml                            6            107            377            291